+++ src/themes/teams/components/Label/labelStyles.ts
@@ -1,6 +1,6 @@
 import * as _ from 'lodash'
 
-import { pxToRem } from '../../utils'
+import { pxToRem } from '../../../../lib'
 import { ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
 import { LabelProps } from '../../../../components/Label/Label'
 import { LabelVariables } from './labelVariables'