+++ src/themes/teams/utils/index.ts
@@ -1,10 +1,5 @@
-import { pxToRem as basePxToRem } from '../../../lib'
 import { ThemeInput, ColorScheme, ColorSchemeMapping } from '../../types'
 import { Partial } from '../../../../types/utils'
-
-const themeFontSizeInPx = 14
-
-export const pxToRem = (sizeInPx: number) => basePxToRem(sizeInPx, themeFontSizeInPx)
 
 export const getSideArrow = (theme: ThemeInput) => {
   const { rtl, siteVariables } = theme